test crawl_bot

+++ crawl_bot.rb
@@ -6,6 +6,7 @@ class CrawlBot
   include Config
 
   def initialize
+    @run_time = rand(14400) + 7200 # random seconds from 2 to 6 hours
     poll
   end
 
@@ -18,44 +19,50 @@ class CrawlBot
         max_number_of_messages: 1,
         visibility_timeout: 10 # keep message invisible long enough to process to wip
       ) do |msg, stats|
-        byebug
         begin
-          # if JSON.parse(msg.body).has_key?('Records')
-          #   puts "\n\nPossible job found:\n\n#{JSON.parse(msg.body)}"
-          #   catch :no_such_job_in_backlog do
-          #     job = 'testing'
-          #     run_job(job)
-          #     puts "finished job:"
-          #   end
-          # else
-          puts "body: #{msg.body}"
-            sqs.delete_message({
+          byebug
+          body = JSON.parse(msg.body)
+          if possible_valid_job?(body)
+            puts "\n\nPossible job found:\n#{body}"
+            catch :no_such_job_in_backlog do
+              job = Job.new(msg, backlog_address)
+              job.run
+              puts "finished job: #{job.params}"
+            end
+          else
+            puts "body: #{msg.body}"
+            sqs.delete_message(
               queue_url: backlog_address,
               receipt_handle: msg.receipt_handle
-            })
-          # end
+            )
+          end
         rescue JSON::ParserError => e
           puts "Trouble with #{msg.body}:\n____#{e}\n"
         end
         puts 'Polling....'
       end
     end
-    ec2.terminate_instancers(ids: [self_id])
+    ec2.terminate_instances(ids: [self_id])
+  end
+
+  def possible_valid_job?(body)
+    !!(body.has_key?('crawler_product_result_id') &&
+        body.has_key?('title'))
   end
 
   def should_stop?
-    if hour_mark_approaches? ? death_ratio_acheived? : false
+    !!(time_is_up? ? death_ratio_acheived? : false)
   end
 
-  def hour_mark_approaches?
-    ((Time.now.to_i - boot_time) % 3600) > 3300
+  def time_is_up?
+    !!(@run_time - Time.now.to_i)
   end
 
   def death_ratio_acheived?
-    death_ratio >= death_threashold
+    !!(death_ratio >= death_threashold)
   end
 
-  def death_ratio #
+  def death_ratio
     backlog = get_count(backlog_address)
     wip = get_count(wip_address)
 
@@ -63,9 +70,10 @@ class CrawlBot
     backlog / wip
   end
 
-  def run_job(job)
-    sleep rand(polling_sleep_time)
-  end
+  # def run_job(job)
+  #   # system("java -jar google-scraper.jar #{job.params.join(' ')}")
+  #   puts "java -jar google-scraper.jar #{job.params.join(' ')}"
+  # end
 
   def self_id # hard code a value here for testing
     @id ||= 'fdasasfd' # HTTParty.get('http://169.254.169.254/latest/meta-data/instance-id')
